+++ src/main/java/de/samply/app/ProjectManagerConst.java
@@ -584,6 +584,8 @@ public class ProjectManagerConst {
     public final static String PRIMARY_MAIL_SENDER = "primaryJavaMailSender";
     public final static String TEST_MAIL_SENDER = "testJavaMailSender";
     public final static String NOT_AUTHORIZED = "Not authorized yet";
+    public final static String FORM_FILENAME_TIMESTAMP_FORMAT = "yyyyMMdd-HHmm";
+    public final static String FORM_FILENAME_TIMESTAMP_ZONE = "UTC";
 
 
 }